requirements.txt
pip==25.0.1
Django==5.0.3
Pillow==10.3.0
django-mptt==0.16.0
django-filter==24.2
webptools==0.0.9
decouple==3.8